treasure_your_passwords
Module encrypt_decrypt_with_ssh_key_mod

Encrypt and decrypt strings with ssh key.

In development, we often use SSH keys to connect to remote computers.
Because of this, we learned how to create and maintain SSH private keys.
This knowledge can be used to encrypt and decrypt many different secrets.
The core principle is simple.

Encrypt

Create a random seed 32 bytes. This is not a secret.
Sign it with the SSH private key. Only the owner can do that.
The signature is our new secret password.
Use symmetrical encryption to encrypt the strings.
Use base64 encoding to convert between strings and bytes.
Save in the same file the encrypted string and the random seed.

Decrypt

Read the file and parse the encrypted string and the random seed.
Sign it with the SSH private key. Only the owner can do that. The signature is our new secret password. Use symmetrical encryption to decrypt the strings.

SSH private key

Private SSH keys are protected by a passcode that only the owner knows.
This is the foundation of the security of this process.
Typing the passcode every time we need the private key is tedious.
We can use the ssh-agent to store the opened private key for some limited time.
This is more convenient, but it is less secure.

Modules

generate_strong_password_mod

Generate strong password.

secret_vault_mod

Secret vault can store many secret tokens.

Structs

EncryptedTextWithMetadata 🔒

Struct that represents the json data saved in the ‘*.enc’ file.

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color

Functions

decode64_from_string_to_32bytes 🔒

Decode base64 from string to 32bytes.

decode64_from_string_to_string 🔒

Decode base64 from string to string.

decrypt_symmetric 🔒

Decrypts plain_encrypted_string with secret_passcode_32bytes.

encode64_from_32bytes_to_string 🔒

Decode base64 from string to 32bytes.

encode64_from_bytes_to_string 🔒

Encode base64 from bytes to string.

encode64_from_string_to_string 🔒

Encode base64 from string to string.

encrypt_symmetric 🔒

Encrypts symmetrically secret_string_to_encrypt with secret_passcode_32bytes.

get_encrypted_file_path 🔒

Get encrypted file path.

get_private_key_file_path 🔒

Get private key file path.

open_file_b64_get_string 🔒

Get the string from the file that is base64 encoded.

random_seed_32bytes_and_string 🔒

Generate a random seed.

shorten_vec_bytes_to_32bytes 🔒

Shorten the Vec<u8> to [u8;32].

sign_seed_with_private_key_file 🔒

Sign the seed with the private key into 32 bytes secret.

sign_seed_with_ssh_agent 🔒

Sign the seed with ssh-agent into 32 bytes secret.

sign_seed_with_ssh_agent_or_private_key_file 🔒

Returns the secret signed seed.