treasure_your_passwords::encrypt_decrypt_with_ssh_key_mod
Module generate_strong_password_mod

Generate strong password.

Starts with a human readable password.
It hashes it to 32 bytes, just to make it a little more complex.
Sign it with the ssh private key, that only the owner can user.
This could be reversed by using the public key. Because of that:
Hash it into a 32 bytes. This is a one way encryption
Finally encode it into a ascii7 string with lowercase, uppercase, numerals and few symbols.

Functions

generate_strong_password 🔒

Generate strong password.